dashboard/requirements.txt
# dashboard/requirements.txt
streamlit==1.29.0
kafka-python==2.0.2
pandas==2.1.4
plotly==5.18.0
minio==7.2.3
